% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advRlab3.R
\name{dijkstra}
\alias{dijkstra}
\title{Dijkstra's algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data frame with three vectors}

\item{init_node}{A start node}
}
\value{
The shortest path from \code{init_node} to the other nodes in \code{graph}
}
\description{
Dijkstra's algorithm
}
\examples{
dijkstra(wiki_graph, 4)

\\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
