\name{my_first_package-package}
\alias{my_first_package-package}
\alias{my_first_package}
\docType{package}
\title{
\packageTitle{my_first_package}
}
\description{
\packageDescription{my_first_package}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{my_first_package}
\packageIndices{my_first_package}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{my_first_package}

Maintainer: \packageMaintainer{my_first_package}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
